requirements.txt
streamlit
pandas
numpy
opencv-python
Pillow
ultralytics